move folder, explicit constructor
